defconfig: Tweak and optimize for LGE_8996.
Our current source builds a lot of drivers and sets many CONFIGs
that aren't actually beneficial to our devices, at the same time
it doesn't enable some settings that are helpful for kernel
functionality and security hardening on MSM8996.

This change should sort all of that out without making the
defconfigs too complicated.