sound: soc: es9218: Fix return values for invalid states and allocs
For some reason, ess' kernel driver returns 0 on many error cases
such as invalid dac states and some allocation failures. This
includes many functions related to sysfs reads and writes, which
rely on return values to give the function caller the amount of
bytes read or written. Returning 0 whenever an error arises on any
of those is problematic since it will point out that no bytes
were written, and not give an indication of the error being hit.