drivers: power: lpm-stats: Fix undefined behavior on strnlen usage
In lpm-stats.c, strnlen is used multiple times in order to return the
number of bytes present on the lpm_stats_reset string. However, the
'maxlen' value passed as a limiter for the max amount of bytes
allowed to be read is "MAX_STR_LEN" (256), which can lead to undefined
behavior since lpm_stats_reset only has 6 chars ("reset" plus a
null terminator), and as such, only 6 bytes of actual data.

Since it has a null terminator at the end, this strnlen call should be
relatively safe as the strnlen call will possibly stop at the terminator.

Still, GCC LTO catches this incorrect usage and throws a warning:

../drivers/power/qcom/lpm-stats.c: In function 'lpm_stats_file_write':
../drivers/power/qcom/lpm-stats.c:265:22: warning: 'strnlen' specified bound 256 exceeds source size 6 [-Wstringop-overread]
  265 |         size_t len = strnlen(lpm_stats_reset, MAX_STR_LEN);

Which should now be fixed.